atm-backend/requirements.txt
fastapi
uvicorn
motor
dnspython
python-dotenv
pydantic
